out_oms_diag: renamed oms_diag.rb to oms_diag_lib.rb
Renamed the file to make it more consistent with the naming
conventions.
