+++ source/code/plugins/oms_diag_lib.rb
