+++ source/code/plugins/out_oms_diag.rb
@@ -12,7 +12,7 @@ module Fluent
       require_relative 'omslog'
       require_relative 'oms_configuration'
       require_relative 'oms_common'
-      require_relative 'oms_diag'
+      require_relative 'oms_diag_lib'
     end
 
     config_param :omsadmin_conf_path, :string, :default => '/etc/opt/microsoft/omsagent/conf/omsadmin.conf'
